Started working on GUI loading.
Need to change the way GUI elements are constructed, though, since there is no
access to the appropriate Render::Stage when the GUI::Loader is executing.

This is also a good time to set up a new rendering system for the GUI instead
of continually creating and destroying Renderables . . . instead using a single
Renderable per widget and updating the geometry/parameters appropriately, hence
removing the dependency upon the Stage object being known in the Widget.